refactor: refactor tests to reduce duplication
